[ci:force] move data-column in datatable
