+++ lib/core/src/lib/datatable/components/datatable/datatable.component.spec.ts
@@ -26,8 +26,8 @@ import { ObjectDataTableAdapter } from '../../data/object-datatable-adapter';
 import { DataTableComponent, ShowHeaderMode } from './datatable.component';
 import { setupTestBed } from '../../../testing/setup-test-bed';
 import { CoreTestingModule } from '../../../testing/core.testing.module';
-import { DataColumnListComponent } from '../../../data-column/data-column-list.component';
-import { DataColumnComponent } from '../../../data-column/data-column.component';
+import { DataColumnListComponent } from '../../data-column/data-column-list.component';
+import { DataColumnComponent } from '../../data-column/data-column.component';
 import { TranslateModule } from '@ngx-translate/core';
 import { domSanitizerMock } from '../../../mock/dom-sanitizer-mock';
 import { matIconRegistryMock } from '../../../mock/mat-icon-registry-mock';
