+++ lib/core/src/lib/datatable/data-column/data-column-list.component.ts
