+++ lib/core/src/lib/datatable/data-column/data-column.component.spec.ts
@@ -16,8 +16,8 @@
  */
 
 import { DataColumnComponent } from './data-column.component';
-import { setupTestBed } from '../testing/setup-test-bed';
-import { CoreTestingModule } from '../testing/core.testing.module';
+import { setupTestBed } from '../../testing/setup-test-bed';
+import { CoreTestingModule } from '../../testing/core.testing.module';
 import { TranslateModule } from '@ngx-translate/core';
 
 describe('DataColumnListComponent', () => {
