+++ lib/core/src/lib/datatable/data-column/data-column.module.ts
@@ -32,9 +32,7 @@ import { DateColumnHeaderComponent } from './data-column-header.component';
         DateColumnHeaderComponent
     ],
     exports: [
-        DataColumnComponent,
-        DataColumnListComponent,
-        DateColumnHeaderComponent
+
     ]
 })
 export class DataColumnModule {}
