+++ lib/core/src/lib/datatable/data/data-table.schema.ts
@@ -18,7 +18,7 @@
 import { ContentChild, Input, Directive } from '@angular/core';
 import { ReplaySubject } from 'rxjs';
 import { AppConfigService } from '../../app-config/app-config.service';
-import { DataColumnListComponent } from '../../data-column/data-column-list.component';
+import { DataColumnListComponent } from '../data-column/data-column-list.component';
 import { DataColumn } from './data-column.model';
 import { ObjectDataColumn } from './object-datacolumn.model';
