+++ lib/core/src/public-api.ts
@@ -23,7 +23,7 @@ export * from './lib/pagination/index';
 export * from './lib/login/index';
 export * from './lib/language-menu/index';
 export * from './lib/info-drawer/index';
-export * from './lib/data-column/index';
+export * from './lib/datatable/data-column/index';
 export * from './lib/datatable/index';
 export * from './lib/context-menu/index';
 export * from './lib/card-view/index';
